LayoutTests/imported/w3c/web-platform-tests/css/css-anchor-position/try-tactic-sizing-expected.txt


PASS flip-block does not affect sizing
PASS flip-inline does not affect sizing
FAIL flip-start affects sizing assert_equals: expected "40px" but got "30px"
